Fixing compile sign warnings -> ints to size_t

+++ proposer.h
@@ -9,26 +9,26 @@
 #include "node.h"
 
 class Proposer{
-private:
-    std::map<int, int> count;
-    
+  private:
+    std::map<int, size_t> count;
+
     size_t quorum;
     int proposal_number;
     std::vector<node> replicas;
-    
-public:
+
+  public:
     /* The value to be proposed */
     std::string to_propose;
-    
+
     void init(vector<node> _replicas);
     Message* start_prepare(int proposal_number);
-    
+
     Message* prepare_accept(int n_a, std::string value);
-    
+
     Message* prepare_reject(int n_p);
-    
+
     Message* propose_accept(int n);
-    
+
     Message* propose_reject(int n_p);
 };
 #endif /* proposer_h */
